@@ -105,53 +105,54 @@
     "connect": "3.7.0",
     "connect-history-api-fallback": "1.6.0",
     "connected-react-router": "6.8.0",
-    "css-loader": "3.4.2",
+    "css-loader": "3.5.3",
     "cssnano": "4.1.10",
-    "current-device": "0.10.0",
-    "date-fns": "2.11.1",
+    "current-device": "0.10.1",
+    "date-fns": "2.14.0",
     "debug": "4.1.1",
-    "ejs": "3.0.2",
+    "ejs": "3.1.3",
     "ejs-loader": "0.3.6",
-    "eslint": "6.8.0",
+    "eslint": "7.1.0",
     "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-react": "7.19.0",
+    "eslint-plugin-react": "7.20.0",
     "eslint-watch": "6.0.1",
     "event-emitter": "0.3.5",
+    "extract-zip": "1.6.6",
     "favicons-webpack-plugin": "3.0.1",
-    "feed": "4.1.0",
+    "feed": "4.2.0",
     "file-loader": "6.0.0",
     "format-number": "3.0.0",
     "git-branch": "2.0.1",
     "history": "4.10.1",
     "iframe-resizer": "4.2.10",
-    "jest": "25.2.4",
-    "jest-cli": "25.2.4",
+    "jest": "26.0.1",
+    "jest-cli": "26.0.1",
     "jest-puppeteer": "4.4.0",
     "jest-standard-reporter": "1.0.4",
-    "js-yaml": "3.13.1",
+    "js-yaml": "3.14.0",
     "jsdom": "16.2.2",
     "json-loader": "0.5.7",
-    "json2csv": "5.0.0",
+    "json2csv": "5.0.1",
     "measure-text-width": "0.0.4",
-    "millify": "3.1.3",
+    "millify": "3.2.1",
     "mini-css-extract-plugin": "0.9.0",
     "minimatch": "3.0.4",
     "node-emoji": "1.10.0",
-    "node-sass": "4.13.1",
+    "node-sass": "4.14.1",
     "postcss-loader": "3.0.0",
     "prop-types": "15.7.2",
-    "puppeteer": "2.1.1",
-    "query-string": "6.11.1",
+    "puppeteer": "3.0.4",
+    "query-string": "6.12.1",
     "raf": "3.4.1",
     "react": "16.13.1",
     "react-chartjs-2": "2.9.0",
     "react-dom": "16.13.1",
     "react-ga": "2.7.0",
-    "react-hot-loader": "4.12.20",
+    "react-hot-loader": "4.12.21",
     "react-key-handler": "1.2.0-beta.3",
     "react-redux": "7.2.0",
-    "react-router": "5.1.2",
-    "react-router-dom": "5.1.2",
+    "react-router": "5.2.0",
+    "react-router-dom": "5.2.0",
     "react-snap": "1.23.0",
     "react-test-renderer": "16.13.1",
     "react-twitter-widgets": "1.7.1",
@@ -167,15 +168,15 @@
     "rimraf": "3.0.2",
     "sass-loader": "8.0.2",
     "serve-static": "1.14.1",
-    "sitemap": "6.1.0",
-    "style-loader": "1.1.3",
-    "svg-autocrop": "2.0.9",
-    "terser-webpack-plugin": "2.3.5",
+    "sitemap": "6.1.4",
+    "style-loader": "1.2.1",
+    "svg-autocrop": "2.0.13",
+    "terser-webpack-plugin": "3.0.1",
     "traverse": "0.6.6",
     "typeface-roboto": "0.0.75",
-    "url-loader": "4.0.0",
-    "webpack": "4.42.1",
-    "webpack-bundle-analyzer": "3.6.1",
+    "url-loader": "4.1.0",
+    "webpack": "4.43.0",
+    "webpack-bundle-analyzer": "3.8.0",
     "webpack-dev-middleware": "3.7.2",
     "webpack-hot-middleware": "2.25.0",
     "webpack-md5-hash": "0.0.6",
