@@ -1,6 +1,6 @@
 import puppeteer from "puppeteer";
 require('expect-puppeteer');
-import devices from 'puppeteer/DeviceDescriptors';
+import { devicesMap } from 'puppeteer/DeviceDescriptors';
 import { paramCase } from 'change-case';
 import { settings } from '../tools/settings';
 import { projects } from '../tools/loadData';
@@ -141,17 +141,17 @@ describe("Normal browser", function() {
 
     console.log(`Checking we see ${project.name} when filtering by organization ${project.organization}`);
     const page = await makePage(`${appUrl}/organization=${organizationSlug}&format=card-mode`);
-    await expect(page).toMatch(project.name);
+    await expect(page).toHaveElement(`//div[contains(@class, 'mosaic')]//*[text()='${project.name}']`);
 
     console.log(`Checking we don't see ${project.name} when filtering by organization ${otherProject.organization}`);
     await page.goto(`${appUrl}/organization=${otherOrganizationSlug}&format=card-mode`);
-    await expect(page).not.toMatch(project.name);
+    await expect(page).not.toHaveElement(`//div[contains(@class, 'mosaic')]//*[text()='${project.name}']`);
   }, 6 * 60 * 1000);
 });
 
 describe("iPhone simulator", function() {
   beforeEach(async function() {
-    setup = async (page) => await page.emulate(devices['iPhone X'])
+    setup = async (page) => await page.emulate(devicesMap['iPhone X'])
   })
 
   afterEach(async function() {
