@@ -1,25 +1,92 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { pure } from 'recompose';
 import LandscapeContent from './LandscapeContent';
+import { calculateSize, outerPadding, headerHeight } from "../../utils/landscapeCalculations";
+import isDesktop from "../../utils/isDesktop";
+
+const calculateZoom = (width, height, zoomedIn) => {
+  const boxHeight = height + headerHeight + 2 * outerPadding
+  const boxWidth = width + 2 * outerPadding
+  const isFirefox = navigator.userAgent.indexOf('Firefox') > -1
+
+  const aspectRatio = innerWidth / innerHeight
+  const adjustedWidth = outerWidth
+  const adjustedHeight = adjustedWidth / aspectRatio
+  let baseZoom = Math.min(adjustedHeight / boxHeight, adjustedWidth / boxWidth).toPrecision(4)
+  let wrapperWidth, wrapperHeight
+
+  if (baseZoom <= 0.95 || !isDesktop || isFirefox || location.search.indexOf('scale=false') > -1) {
+    wrapperWidth = Math.max(boxWidth, innerWidth)
+    wrapperHeight = Math.max(boxHeight, innerHeight)
+    baseZoom = 1
+  } else {
+    wrapperWidth = adjustedWidth / baseZoom
+    wrapperHeight = adjustedHeight / baseZoom
+  }
+
+  return { zoom: Math.min(baseZoom * (zoomedIn ? 3 : 1), 3), wrapperWidth, wrapperHeight }
+}
 
 const Fullscreen = ({ready, groupedItems, landscapeSettings, version}) => {
   if (ready !== true) {
-    return (
-      <div>
-      </div>
-    )
+    return <div></div>
   }
 
+  const [_, setWindowSize] = useState(1)
+  const [zoomedIn, setZoomedIn] = useState(false)
+  const [zoomedAt, setZoomedAt] = useState({})
+  const onZoom = (e) => {
+    setZoomedAt({ x: e.pageX / zoom, y: e.pageY / zoom })
+    setZoomedIn(!zoomedIn)
+  }
+
+  useEffect(() => {
+    const calculateWindowSize = () => setWindowSize(`${innerWidth}x${innerHeight}`)
+    window.addEventListener("resize", calculateWindowSize)
+    return () => window.removeEventListener("resize", calculateWindowSize)
+  }, [true]);
+
+  useEffect(() => {
+    zoomedIn ? window.scrollTo((zoomedAt.x * zoom - innerWidth / 2), (zoomedAt.y * zoom - innerHeight / 2)) : null
+  }, [zoomedAt, zoomedIn])
+
+  const { width, height } = calculateSize(landscapeSettings)
+  const { zoom, wrapperWidth, wrapperHeight } = calculateZoom(width, height, zoomedIn)
+
   return (
-    <div style={{zoom: 4, fontFamily: 'roboto'}}>
-        <div className="gradient-bg" style={{
-          width: landscapeSettings.fullscreen_size.width,
-          height: landscapeSettings.fullscreen_size.height,
-          position: 'relative'}}>
-          <LandscapeContent style={{top: 50, left: 20}} groupedItems={groupedItems} zoom={1} landscapeSettings={landscapeSettings} />
+      <div className="gradient-bg" style={{
+        fontFamily: 'roboto',
+        width: wrapperWidth * zoom,
+        height: wrapperHeight * zoom,
+        overflow: 'hidden'
+        }}>
+        <div style={{
+          transform: `scale(${zoom})`,
+          width: wrapperWidth,
+          height: wrapperHeight,
+          transformOrigin: '0 0',
+          paddingTop: headerHeight,
+          position: 'relative',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center'
+        }}>
+          <div
+            onClick={isDesktop && onZoom}
+            style={{
+              position: 'absolute',
+              top: 0,
+              left: 0,
+              right: 0,
+              bottom: 0,
+              cursor: zoomedIn ? 'zoom-out' : 'zoom-in',
+              zIndex: 100000
+            }}>
+          </div>
+          <LandscapeContent groupedItems={groupedItems} landscapeSettings={landscapeSettings} padding={0} />
           <div style={{
             position: 'absolute',
-            top: 15,
+            top: 10,
             left: '50%',
             transform: 'translateX(-50%)',
             fontSize: 18,
@@ -35,7 +102,7 @@ const Fullscreen = ({ready, groupedItems, landscapeSettings, version}) => {
             position: 'absolute',
             top: 15,
             right: 12,
-            fontSize: 6,
+            fontSize: 11,
             background: '#eee',
             color: 'rgb(100,100,100)',
             paddingLeft: 20,
@@ -59,7 +126,7 @@ const Fullscreen = ({ready, groupedItems, landscapeSettings, version}) => {
             color: '#eee',
           }}>{version}</div>
         </div>
-    </div>
+      </div>
   );
 };
 
