@@ -2,34 +2,36 @@ import React from 'react';
 import { pure } from 'recompose';
 import _ from 'lodash';
 
-const LandscapeInfo = ({zoom, width, height, top, left, childrenInfo}) => {
+const LandscapeInfo = ({width, height, top, left, childrenInfo}) => {
   const children = childrenInfo.map(function(info) {
     const positionProps = {
         position: 'absolute',
-        top: _.isUndefined(info.top) ? null : info.top * zoom,
-        left: _.isUndefined(info.left) ? null : info.left * zoom,
-        right: _.isUndefined(info.right) ? null : info.right * zoom,
-        bottom: _.isUndefined(info.bottom) ? null : info.bottom * zoom ,
-        width: _.isUndefined(info.width) ? null : info.width * zoom,
-        height: _.isUndefined(info.height) ? null : info.height * zoom
+        top: _.isUndefined(info.top) ? null : info.top,
+        left: _.isUndefined(info.left) ? null : info.left,
+        right: _.isUndefined(info.right) ? null : info.right,
+        bottom: _.isUndefined(info.bottom) ? null : info.bottom,
+        width: _.isUndefined(info.width) ? null : info.width,
+        height: _.isUndefined(info.height) ? null : info.height
     };
     if (info.type === 'text') {
       // pdf requires a normal version without a zoom trick
       if (window.location.href.indexOf('&pdf') !== -1) {
         return <div key='text' style={{
           ...positionProps,
-          fontSize: info.font_size * zoom,
+          fontSize: info.font_size,
           fontStyle: 'italic',
-          textAlign: 'justify'
+          textAlign: 'justify',
+          zIndex: 1
         }}>{info.text}</div>
       // while in a browser we use a special version which renders fonts
       // properly on a small zoom
       } else {
         return <div key='text' style={{
           ...positionProps,
-          fontSize: info.font_size * zoom * 4,
+          fontSize: info.font_size * 4,
           fontStyle: 'italic',
-          textAlign: 'justify'
+          textAlign: 'justify',
+          zIndex: 1
         }}><div style={{
           position: 'absolute',
           left: 0,
@@ -46,7 +48,7 @@ const LandscapeInfo = ({zoom, width, height, top, left, childrenInfo}) => {
     if (info.type === 'title') {
       return <div key='title' style= {{
         ...positionProps,
-        fontSize: info.font_size * zoom,
+        fontSize: info.font_size,
         color: '#666'
       }}>{info.title}</div>
     }
@@ -57,15 +59,15 @@ const LandscapeInfo = ({zoom, width, height, top, left, childrenInfo}) => {
 
   return <div style={{
     position: 'absolute',
-    width: width * zoom,
-    height: (height - 20) * zoom,
-    top: top * zoom,
-    left: left * zoom,
-    border: `${1 * zoom}px solid black`,
+    width: width,
+    height: height - 20,
+    top: top,
+    left: left,
+    border: '1px solid black',
     background: 'white',
-    borderRadius: 15 * zoom,
-    marginTop: 20 * zoom,
-    boxShadow: `0 ${4 * zoom}px ${8 * zoom}px 0 rgba(0, 0, 0, 0.2), 0 ${6 * zoom}px ${20 * zoom}px 0 rgba(0, 0, 0, 0.19)`
+    borderRadius: 10,
+    marginTop: 20,
+    boxShadow: `0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19)`
   }}>{children}</div>
 }
 export default pure(LandscapeInfo);
