@@ -1,9 +1,9 @@
-import Promise from 'bluebird';
 import { env } from 'process';
 import { stringify } from 'query-string';
 import rp from './rpRetry';
+import _ from 'lodash'
 
-['CRUNCHBASE_KEY', 'GITHUB_KEY', 'TWITTER_KEYS'].forEach((key) => {
+['CRUNCHBASE_KEY_4', 'GITHUB_KEY', 'TWITTER_KEYS'].forEach((key) => {
   if (!env[key]) {
     console.info(`${key} not provided`);
   }
@@ -12,7 +12,7 @@ import rp from './rpRetry';
 let requests = {};
 
 // We only want to retry a request when rate limited. By default the status code is 429.
-const ApiClient = ({ baseUrl, defaultOptions = {}, defaultParams = {}, retryStatuses = [429] }) => {
+const ApiClient = ({ baseUrl, defaultOptions = {}, defaultParams = {}, retryStatuses = [429], delayFn = null }) => {
   return {
     request: async ({ path = null, url = null, method = 'GET', params = {} }) => {
       const qs = { ...defaultParams, ...params };
@@ -30,7 +30,8 @@ const ApiClient = ({ baseUrl, defaultOptions = {}, defaultParams = {}, retryStat
           json: true,
           ...defaultOptions,
           qs,
-          retryStatuses
+          retryStatuses,
+          delayFn
         })
       }
 
@@ -40,14 +41,25 @@ const ApiClient = ({ baseUrl, defaultOptions = {}, defaultParams = {}, retryStat
 };
 
 export const CrunchbaseClient = ApiClient({
-  baseUrl: 'https://api.crunchbase.com/v3.1',
-  defaultParams: { user_key: env.CRUNCHBASE_KEY },
+  baseUrl: 'https://api.crunchbase.com/api/v4',
+  defaultParams: { user_key: env.CRUNCHBASE_KEY_4 },
   defaultOptions: { followRedirect: true, maxRedirects: 5, timeout: 10 * 1000 }
 });
 
-const OldGithubClient = ApiClient({
+export const GithubClient = ApiClient({
   baseUrl: 'https://api.github.com',
   retryStatuses: [403], // Github returns 403 when rate limiting.
+  delayFn: error => {
+    const rateLimitRemaining = parseInt(_.get(error, ['response', 'headers', 'x-ratelimit-remaining'], 1))
+    const rateLimitReset = parseInt(_.get(error, ['response', 'headers', 'x-ratelimit-reset'], 1)) * 1000
+    if (rateLimitRemaining > 0) {
+      return 30000
+    } else {
+      const delay = Math.ceil((new Date(rateLimitReset)) - (new Date()))
+      console.log(`Hourly rate limit exceeded on Github, delaying for ${Math.round(delay / 1000 / 60)} minutes`)
+      return delay
+    }
+  },
   defaultOptions: {
     followRedirect: true,
     timeout: 10 * 1000,
@@ -57,24 +69,6 @@ const OldGithubClient = ApiClient({
     },
   }
 });
-
-export const GithubClient = {
-  request: async function({ path = null, url = null, method = 'GET', params = {} }) {
-    const rates = await rp({
-      uri: 'https://api.github.com/rate_limit',
-      json: true,
-      headers: {
-        'User-agent': 'CNCF',
-        'Authorization': `token ${env.GITHUB_KEY}`
-      }});
-    const remaining = rates.resources.core.remaining;
-    if (remaining < 100) { // because requests may go in parallel
-      console.info('Pausing for one hour, github quote exceeded');
-      await Promise.delay(3600 * 1000);
-    }
-    return await OldGithubClient.request({path, url, method, params});
-  }
-};
 
 const [consumerKey, consumerSecret, accessTokenKey, accessTokenSecret] = (env.TWITTER_KEYS || '').split(',');
 
@@ -88,4 +82,8 @@ export const TwitterClient = ApiClient({
       access_token_secret: accessTokenSecret
     }
   }
+});
+
+export const YahooFinanceClient = ApiClient({
+  baseUrl: 'https://query2.finance.yahoo.com',
 });
