@@ -1,6 +1,5 @@
 import { setFatalError } from './fatalErrors';
 import colors from 'colors';
-import rp from './rpRetry'
 import Promise from 'bluebird'
 import _ from 'lodash';
 import ensureHttps from './ensureHttps';
@@ -12,7 +11,7 @@ const error = colors.red;
 const fatal = (x) => colors.red(colors.inverse(x));
 const cacheMiss = colors.green;
 const debug = require('debug')('cb');
-import { CrunchbaseClient } from './apiClients';
+import { CrunchbaseClient, YahooFinanceClient } from './apiClients';
 
 export async function getCrunchbaseOrganizationsList() {
   const traverse = require('traverse');
@@ -62,42 +61,27 @@ export async function extractSavedCrunchbaseEntries() {
   return _.uniq(organizations);
 }
 
-async function getParentCompanies(companyInfo, visited = []) {
-  var parentInfo = companyInfo.relationships.owned_by.item;
-  // console.info(`looking for parent for ${companyInfo.name}`);
-  // console.info(`parentInfo is ${parentInfo}`);
-  if (!parentInfo) {
-    // console.info('returning []');
-    return [];
-  } else {
-    var parentId = parentInfo.uuid;
-    const { permalink } = parentInfo.properties;
-    if (parentId === companyInfo.uuid) {
-      return []; //we are the parent and this hangs up the algorythm
-    }
-    if (visited.includes(permalink)) {
-      throw new Error(`Cyclic dependency detected when fetching parents: ${visited.join(', ')}, ${permalink}`);
-    }
-    var fullParentInfo = await CrunchbaseClient.request({ path: `/organizations/${parentId}` });
-    var cbInfo = fullParentInfo.data;
-    return [parentInfo].concat(await getParentCompanies(cbInfo, [...visited, permalink]));
-  }
+const getCurrencyConversion = async currency => {
+  const path = `/v10/finance/quoteSummary/${currency}=X?modules=summaryDetail`
+  const response = await YahooFinanceClient.request({ path })
+  const { bid, ask } = response.quoteSummary.result[0].summaryDetail
+  return (bid.raw + ask.raw) / 2
 }
+
+const getYahooFinanceData = async ticker => {
+  const path = `/v10/finance/quoteSummary/${ticker}?modules=summaryDetail`
+  const response = await YahooFinanceClient.request({ path })
+  const { currency, marketCap } = response.quoteSummary.result[0].summaryDetail
+  const conversion = currency === 'USD' ? 1 : await getCurrencyConversion(currency)
+  return { ...marketCap, raw: Math.round(marketCap.raw / conversion) }
+}
+
 const marketCapCache = {};
 async function getMarketCap(ticker) {
-  // console.info(ticker, stock_exchange);
-  // console.info('ticker is', ticker);
   debug(`Extracting the ticker from ${ticker}`);
-  var quote;
+  let quote;
   try {
-    quote = marketCapCache[ticker];
-    if (!quote) {
-      const response = (await rp({
-        url: `https://query2.finance.yahoo.com/v10/finance/quoteSummary/${ticker}?modules=summaryDetail`,
-        json: true
-      }));
-      quote = response.quoteSummary.result[0].summaryDetail.marketCap;
-    }
+    quote = marketCapCache[ticker] || await getYahooFinanceData(ticker)
   } catch(ex) {
     throw new Error(`Can't resolve stock ticker ${ticker}; please manually add a "stock_ticker" key to landscape.yml or set to null`);
   }
@@ -110,31 +94,84 @@ async function getMarketCap(ticker) {
   return result;
 }
 
-const fetchAcquisitions = async(url, params = {}) => {
-  const response = await CrunchbaseClient.request({ url, params })
-  const { items, paging } = response.data
-  const { next_page_url } = paging
-  if (!next_page_url) {
-    return items
-  }
-  return items + await fetchAcquisitions(next_page_url)
-}
+export async function fetchData(name) {
+  const result = await CrunchbaseClient.request({
+    path: `entities/organizations/${name}`,
+    params:{'card_ids': 'headquarters_address,acquiree_acquisitions,parent_organization', 'field_ids': 'num_employees_enum,linkedin,twitter,name,website,short_description,funding_total,stock_symbol,stock_exchange_symbol' }
+  });
 
-const getAcquisitions = async (acquisitions) => {
-  const { items, paging } = acquisitions
-  const { total_items, first_page_url } = paging
-  const entries = items.length === total_items ? items : await fetchAcquisitions(first_page_url, { items_per_page: 1000 })
-  return entries.map(acquisition => {
-    const { name } = acquisition.relationships.acquiree.properties
-    let result = {
-      date: acquisition.properties.announced_on,
-      acquiree: name
+  const mapAcquisitions = function(a) {
+    const result = {
+      date: a.announced_on.value,
+      acquiree: a.acquiree_identifier.value,
     }
-    if (acquisition.properties.price_usd) {
-      result.price = acquisition.properties.price_usd
+    if (a.price) {
+      result.price = a.price.value_usd
     }
-    return result
-  })
+    return result;
+  }
+  let acquisitions = result.cards.acquiree_acquisitions.map(mapAcquisitions);
+  const limit = 100;
+  let lastPage = result;
+  while (lastPage.cards.acquiree_acquisitions.length === limit) {
+     lastPage = await CrunchbaseClient.request({
+        path: `entities/organizations/${name}/cards/acquiree_acquisitions`,
+        params: {after_id: lastPage.cards.acquiree_acquisitions[limit - 1].identifier.uuid}
+      });
+      acquisitions = acquisitions.concat(lastPage.cards.acquiree_acquisitions.map(mapAcquisitions));
+  }
+  acquisitions = _.orderBy(acquisitions, ['date', 'acquiree']);
+
+  let parents = [];
+  let lastOrganization = result;
+  while (lastOrganization.cards.parent_organization[0]) {
+    parents.push(lastOrganization.cards.parent_organization[0]);
+    const url =  `entities/organizations/${lastOrganization.cards.parent_organization[0].identifier.permalink}`;
+    lastOrganization = await CrunchbaseClient.request({
+      path: url,
+      params:{'card_ids': 'parent_organization', 'field_ids': '' }
+    });
+  }
+  const parentLinks = parents.map( (item) => 'https://www.crunchbase.com/organization/' + item.identifier.permalink );
+
+  const firstWithStockSymbol = _.find([result.properties].concat(parents), (x) => !!x.stock_symbol);
+  const stockSymbol = firstWithStockSymbol ? firstWithStockSymbol.stock_symbol.value : undefined;
+  const firstWithTotalFunding = _.find([result.properties].concat(parents), (x) => !!x.funding_total);
+  const totalFunding = firstWithTotalFunding ? + firstWithTotalFunding.funding_total.value_usd.toFixed() : undefined;
+
+  const getAddressPart = function(part) {
+    return (result.cards.headquarters_address[0].location_identifiers.filter( (x) => x.location_type === part)[0] || {}).value
+  }
+
+
+  const { employeesMin, employeesMax } = (function() {
+    const value = result.properties.num_employees_enum || '';
+    const parts = value.split('_');
+    if (parts.length !== 3) {
+      return { employeesMin: null, employeesMax: null}
+    } else {
+      return { employeesMin: + parts[1], employeesMax: parts[2] === 'max' ? 1000000 : + parts[2] }
+    }
+  })();
+
+  const employee_parts = (result.properties.num_employees_enum || '').split('_');
+  return {
+    name: result.properties.name,
+    description: result.properties.short_description,
+    num_employees_min: employeesMin,
+    num_employees_max: employeesMax,
+    homepage: (result.properties.website || {value: null}).value ,
+    city: getAddressPart('city'),
+    region: getAddressPart('region'),
+    country: getAddressPart('country'),
+    twitter: result.properties.twitter ? result.properties.twitter.value : null,
+    linkedin: result.properties.linkedin ? ensureHttps(result.properties.linkedin.value) : null,
+    acquisitions,
+    parents: parentLinks,
+    ticker: stockSymbol,
+    funding: totalFunding,
+    stockExchange: result.properties.stock_exchange_symbol || null
+  }
 }
 
 export async function fetchCrunchbaseEntries({cache, preferCache}) {
@@ -155,26 +192,11 @@ export async function fetchCrunchbaseEntries({cache, preferCache}) {
       return {};
     }
     try {
-      const result = await CrunchbaseClient.request({ path: `/organizations/${c.name}` });
-      var cbInfo = result.data.properties;
-      const { relationships } = result.data
-      var twitterEntry = _.find(relationships.websites.items, (x) => x.properties.website_name === 'twitter');
-      var linkedInEntry = _.find(relationships.websites.items, (x) => x.properties.website_name === 'linkedin');
-      const headquarters = relationships.headquarters;
-      const acquisitions = await getAcquisitions(relationships.acquisitions)
+      const result = await fetchData(c.name);
+
       const entry = {
         url: c.crunchbase,
-        name: cbInfo.name,
-        description: cbInfo.short_description,
-        num_employees_min: cbInfo.num_employees_min,
-        num_employees_max: cbInfo.num_employees_max,
-        homepage: cbInfo.homepage_url,
-        city: headquarters && headquarters.item && headquarters.item.properties.city || null,
-        region: headquarters && headquarters.item && headquarters.item.properties.region || null,
-        country: headquarters && headquarters.item && headquarters.item.properties.country || null,
-        twitter: twitterEntry ? twitterEntry.properties.url : null,
-        linkedin: linkedInEntry ? ensureHttps(linkedInEntry.properties.url) : null,
-        acquisitions
+        ...result
       };
       if (_.isEmpty(entry.city)) {
         addError('crunchbase');
@@ -184,26 +206,17 @@ export async function fetchCrunchbaseEntries({cache, preferCache}) {
         reporter.write(fatal("F"));
         return null;
       }
-      var parents = await getParentCompanies(result.data);
-      entry.parents = parents.map( (item) =>  'https://www.crunchbase.com/' + item.properties.web_path);
-       // console.info(parents.map( (x) => x.properties.name));
-      var meAndParents = [result.data].concat(parents);
-      var firstWithTicker = _.find( meAndParents, (org) => !!org.properties.stock_symbol );
-      var firstWithFunding = _.find( meAndParents, (org) => !!org.properties.total_funding_usd );
-      if (!(c.ticker === null) && (firstWithTicker || c.ticker)) {
+
+      if (!(c.ticker === null) && (entry.ticker || c.ticker)) {
         // console.info('need to get a ticker?');
-        entry.ticker = firstWithTicker ? firstWithTicker.properties.stock_symbol : undefined;
         entry.effective_ticker = c.ticker || entry.ticker;
-        entry.market_cap = await getMarketCap(entry.effective_ticker, cbInfo.stock_exchange);
+        entry.market_cap = await getMarketCap(entry.effective_ticker, entry.stockExchange);
         entry.kind = 'market_cap';
-        // console.info(cbInfo.name, 'ticker: ', entry.ticker, ' market cap: ', entry.funding);
-      } else if (firstWithFunding) {
+      } else if (entry.funding) {
         entry.kind = 'funding';
-        entry.funding = firstWithFunding.properties.total_funding_usd;
-        // console.info(cbInfo.name, 'funding: ', entry.funding);
       } else {
-        // console.info(cbInfo.name, 'no finance info');
       }
+      delete entry.stockSymbol;
       reporter.write(cacheMiss("*"));
       return entry;
       // console.info(entry);
