@@ -1,20 +1,18 @@
 import { setFatalError } from './fatalErrors';
 import colors from 'colors';
-import path from 'path';
-import fs from 'fs';
 const Promise = require('bluebird');
-const traverse = require('traverse');
 import _ from 'lodash';
 import rp from './rpRetry';
 import { JSDOM } from 'jsdom';
 import { addError, addWarning } from './reporter';
-import { settings, projectPath } from './settings';
 import makeReporter from './progressReporter';
 const debug = require('debug')('github');
 import shortRepoName from '../src/utils/shortRepoName';
 import getRepositoryInfo , { getLanguages, getWeeklyContributions}  from './getRepositoryInfo';
+import { cacheKey, getRepos } from './repos'
+import { GithubClient} from './apiClients'
 
-import { getRepoLatestDate ,getReleaseDate } from './githubDates';
+import { getRepoLatestDate, getReleaseDate } from './githubDates';
 
 const githubColors = JSON.parse(require('fs').readFileSync('tools/githubColors.json', 'utf-8'));
 
@@ -22,78 +20,47 @@ const error = colors.red;
 const fatal = (x) => colors.red(colors.inverse(x));
 const cacheMiss = colors.green;
 
-export async function extractSavedGithubEntries() {
-  const result = [];
-  const traverse = require('traverse');
-  let source = [];
-  try {
-    source =  require('js-yaml').safeLoad(fs.readFileSync(path.resolve(projectPath, 'processed_landscape.yml')));
-  } catch(_ex) {
-    console.info('Can not extract github entries from the processed_landscape.yml');
-  }
-  const tree = traverse(source);
-  tree.map(function(node) {
-    if (!node) {
-      return;
-    }
-    if (node.github_data) {
-      result.push({...node.github_data, url: node.repo_url, branch: node.branch});
-    }
+const getContributorsCount = async (repoUrl) => {
+  var response = await rp({
+    uri: `${repoUrl}/contributors_size`,
+    followRedirect: true,
+    timeout: 30 * 1000,
+    simple: true
   });
-  return result;
-}
+  const dom = new JSDOM(response);
+  const doc = dom.window.document;
+  var element = doc.querySelector('.num');
+  var count = element.textContent.replace(/[^\d]/g, '').trim();
+  return parseInt(count);
+};
 
-async function getGithubRepos() {
-  const source =  require('js-yaml').safeLoad(fs.readFileSync(path.resolve(projectPath, 'landscape.yml')));
-  const tree = traverse(source);
-  const repos = [];
-  tree.map(function(node) {
-    if (!node) {
-      return;
-    }
-    if (node.item !== null) {
-      return;
-    }
-    if (node.repo_url && node.repo_url.indexOf('https://github.com') === 0) {
-      repos.push({
-        url: node.repo_url,
-        branch: node.branch
-      });
-    } /* else {
-      if (!node.repo_url) {
-        console.info(`item: ${node.name} has no repo url`)
-      } else {
-        console.info(`item: ${node.name} has a non github repo url`)
-      }
-    } */
-  });
-  return _.uniq(repos);
-}
+const getContributorsList = async (repo, page = 1) => {
+  const per_page = 100 // maximum allowed by Github
+  const response = await GithubClient.request({ path: `/repos/${repo}/contributors`, params: { page, per_page }})
+  const contributors = response.map(contributor => contributor.login)
 
+  return contributors.concat(contributors.length === per_page ? await getContributorsList(repo, page + 1) : [])
+}
 
 export async function fetchGithubEntries({cache, preferCache}) {
-  const repos = await getGithubRepos();
+  const repos = getRepos();
   debug(cache);
   const errors = [];
   const reporter = makeReporter();
   const result = await Promise.map(repos, async function(repo) {
-    const cachedEntry = _.find(cache, {url: repo.url, branch: repo.branch});
+    const cachedEntry = cache[cacheKey(repo.url, repo.branch)]
     if (cachedEntry && preferCache) {
       debug(`Cache ${cachedEntry} found for ${repo.url}`);
       reporter.write('.');
-      return cachedEntry;
+      return { ...cachedEntry, cached: true };
+    }
+    if (preferCache && repo.parent && cache[repo.parent]) {
+      return {}
     }
     debug(`No cache found for ${repo.url} ${repo.branch}`);
     await Promise.delay(1 * 1000);
     try {
       const url = repo.url;
-      if (url.split('/').length !==  5 || !url.split('/')[4]) {
-        addError('github');
-        setFatalError(`${repo.url} does not look like a github repo`);
-        reporter.write(fatal('F'));
-        errors.push(fatal(`${repo.url} does not look like a github repo`));
-        return null;
-      }
       const repoName = shortRepoName(url);
       const apiInfo = await getRepositoryInfo(url);
       const languagesInfo = await getLanguages(url);
@@ -119,20 +86,8 @@ export async function fetchGithubEntries({cache, preferCache}) {
 
       const releaseDate = await getReleaseDate({repo: repoName});
       const releaseLink = releaseDate && `${url}/releases`;
-      const getContributorsCount = async function() {
-        var response = await rp({
-          uri: `${url}/contributors_size`,
-          followRedirect: true,
-          timeout: 30 * 1000,
-          simple: true
-        });
-        const dom = new JSDOM(response);
-        const doc = dom.window.document;
-        var element = doc.querySelector('.num');
-        var count = element.textContent.replace(/[^\d]/g, '').trim();
-        return parseInt(count);
-      };
-      const contributorsCount = await getContributorsCount();
+      const contributorsCount = await getContributorsCount(url);
+      const contributorsList = repo.multiple ? await getContributorsList(repoName) : []
       const contributorsLink = `${url}/graphs/contributors`;
       // console.info(contributorsCount, contributorsLink);
       var date;
@@ -155,7 +110,8 @@ export async function fetchGithubEntries({cache, preferCache}) {
         release_date: releaseDate,
         release_link: releaseLink,
         contributors_count: contributorsCount,
-        contributors_link: contributorsLink
+        contributors_link: contributorsLink,
+        contributors_list: contributorsList
       });
     } catch (ex) {
       debug(`Fetch failed for ${repo.url}, attempt to use a cached entry`);
