@@ -1,4 +1,4 @@
-import updateProcessedLandscape from "./updateProcessedLandscape";
+import { updateProcessedLandscape } from "./processedLandscape";
 
 updateProcessedLandscape((processedLandscape) => {
   const count = process.argv[2] ? parseInt(process.argv[2]) : 0;
